Convert Auth* components to TypeScript
I've also gone to the liberty of sneaking in some functional component
goodness.
