+++ src/components/views/auth/AuthBody.tsx
@@ -1,7 +1,5 @@
 /*
-Copyright 2015, 2016 OpenMarket Ltd
 Copyright 2019 New Vector Ltd
-Copyright 2019 The Matrix.org Foundation C.I.C.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,15 +14,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { _t } from '../../../languageHandler';
+'use strict';
+
 import React from 'react';
 
-export default class AuthFooter extends React.Component {
-    render() {
-        return (
-            <div className="mx_AuthFooter">
-                <a href="https://matrix.org" target="_blank" rel="noreferrer noopener">{ _t("powered by Matrix") }</a>
-            </div>
-        );
-    }
+interface IProps {
+    children?: React.ReactNode;
+}
+
+export default function AuthBody (props: IProps) {
+    return (
+        <div className="mx_AuthBody">
+            { props.children }
+        </div>
+    );
 }
